refactor: split chat history text helpers
